review: rename datetime serializer to datetime_rfc3339
