--- src/v2_0_1/helpers/serializer.rs
@@ -1,20 +1,0 @@
-pub mod datetime {
-    use chrono::{DateTime, SecondsFormat, Utc};
-    use serde::{Deserialize, Deserializer, Serializer};
-    pub fn serialize<S>(date: &DateTime<Utc>, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        serializer.serialize_str(&date.to_rfc3339_opts(SecondsFormat::Millis, true))
-    }
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<DateTime<Utc>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s: String = Deserialize::deserialize(deserializer)?;
-        DateTime::parse_from_rfc3339(&s)
-            .map(|dt| dt.with_timezone(&Utc))
-            .map_err(serde::de::Error::custom)
-    }
-}
