+++ src/v2_0_1/messages/heartbeat.rs
@@ -3,7 +3,7 @@
 use chrono::DateTime;
 use chrono::Utc;
 
-use crate::v2_0_1::helpers::serializer::datetime;
+use crate::v2_0_1::helpers::datetime_rfc3339;
 
 /// HeartbeatRequest, sent by the Charging Station to the CSMS. No fields are defined.
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
@@ -17,6 +17,6 @@ pub struct HeartbeatRequest {
 #[serde(rename_all = "camelCase")]
 pub struct HeartbeatResponse {
     /// Contains the current time of the CSMS.
-    #[serde(with = "datetime")]
+    #[serde(with = "datetime_rfc3339 ")]
     pub current_time: DateTime<Utc>,
 }
