+++ src/v2_0_1/messages/notify_customer_information.rs
@@ -1,7 +1,7 @@
 use chrono::DateTime;
 use chrono::Utc;
 
-use crate::v2_0_1::helpers::serializer::datetime;
+use crate::v2_0_1::helpers::datetime_rfc3339;
 
 /// This contains the field definition of the NotifyCustomerInformationRequest PDU sent by the Charging Station to the CSMS.
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
@@ -11,7 +11,7 @@ pub struct NotifyCustomerInformationRequest {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub tbc: Option<bool>,
     pub seq_no: i32,
-    #[serde(with = "datetime")]
+    #[serde(with = "datetime_rfc3339 ")]
     pub generated_at: DateTime<Utc>,
     pub request_id: i32,
 }
