+++ src/v2_0_1/messages/notify_ev_charging_schedule.rs
@@ -4,13 +4,13 @@ use chrono::Utc;
 use crate::v2_0_1::datatypes::charging_schedule_type::ChargingScheduleType;
 use crate::v2_0_1::datatypes::status_info_type::StatusInfoType;
 use crate::v2_0_1::enumerations::generic_status_enum_type::GenericStatusEnumType;
-use crate::v2_0_1::helpers::serializer::datetime;
+use crate::v2_0_1::helpers::datetime_rfc3339;
 
 /// The Charging Station uses this message to communicate the charging needs as calculated by the EV to the CSMS.
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct NotifyEVChargingScheduleRequest {
-    #[serde(with = "datetime")]
+    #[serde(with = "datetime_rfc3339 ")]
     pub time_base: DateTime<Utc>,
     pub evse_id: i32,
     pub charging_schedule: ChargingScheduleType,
