+++ src/v2_0_1/messages/notify_event.rs
@@ -2,13 +2,13 @@ use chrono::DateTime;
 use chrono::Utc;
 
 use crate::v2_0_1::datatypes::event_data_type::EventDataType;
-use crate::v2_0_1::helpers::serializer::datetime;
+use crate::v2_0_1::helpers::datetime_rfc3339;
 
 /// This contains the field definition of the NotifyEventRequest PDU sent by the Charging Station to the CSMS.
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct NotifyEventRequest {
-    #[serde(with = "datetime")]
+    #[serde(with = "datetime_rfc3339 ")]
     pub generated_at: DateTime<Utc>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub tbc: Option<bool>,
