+++ src/v2_0_1/messages/notify_report.rs
@@ -2,7 +2,7 @@ use chrono::DateTime;
 use chrono::Utc;
 
 use crate::v2_0_1::datatypes::report_data_type::ReportDataType;
-use crate::v2_0_1::helpers::serializer::datetime;
+use crate::v2_0_1::helpers::datetime_rfc3339;
 
 /// This contains the field definition of the NotifyReportRequest PDU sent by the Charging Station to the CSMS.
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
@@ -12,7 +12,7 @@ pub struct NotifyReportRequest {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub tbc: Option<bool>,
     pub seq_no: i32,
-    #[serde(with = "datetime")]
+    #[serde(with = "datetime_rfc3339 ")]
     pub generated_at: DateTime<Utc>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub report_data: Option<Vec<ReportDataType>>,
