+++ src/v2_0_1/messages/reserve_now.rs
@@ -5,14 +5,14 @@ use crate::v2_0_1::datatypes::id_token_type::IdTokenType;
 use crate::v2_0_1::datatypes::status_info_type::StatusInfoType;
 use crate::v2_0_1::enumerations::connector_enum_type::ConnectorEnumType;
 use crate::v2_0_1::enumerations::reserve_now_status_enum_type::ReserveNowStatusEnumType;
-use crate::v2_0_1::helpers::serializer::datetime;
+use crate::v2_0_1::helpers::datetime_rfc3339;
 
 /// This contains the field definitions of the RequestStopTransactionRequest PDU sent to Charging Station by CSMS.
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct ReserveNowRequest {
     pub id: i32,
-    #[serde(with = "datetime")]
+    #[serde(with = "datetime_rfc3339 ")]
     pub expiry_date_time: DateTime<Utc>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub connector_type: Option<ConnectorEnumType>,
