+++ src/v2_0_1/messages/transaction_event.rs
@@ -10,14 +10,14 @@ use crate::v2_0_1::datatypes::meter_value_type::MeterValueType;
 use crate::v2_0_1::datatypes::transaction_type::TransactionType;
 use crate::v2_0_1::enumerations::transaction_event_enum_type::TransactionEventEnumType;
 use crate::v2_0_1::enumerations::trigger_reason_enum_type::TriggerReasonEnumType;
-use crate::v2_0_1::helpers::serializer::datetime;
+use crate::v2_0_1::helpers::datetime_rfc3339;
 
 /// Sent by the Charging Station to the CSMS to request that the Certificate Authority signs the public key into a certificate.
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct TransactionEventRequest {
     pub event_type: TransactionEventEnumType,
-    #[serde(with = "datetime")]
+    #[serde(with = "datetime_rfc3339")]
     pub timestamp: DateTime<Utc>,
     pub trigger_reason: TriggerReasonEnumType,
     pub seq_no: i32,
